requirements.txt
aiofiles==24.1.0
aiogram==3.22.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.11.0
async-timeout==5.0.1
attrs==25.3.0
blinker==1.9.0
certifi==2025.8.3
click==8.1.8
exceptiongroup==1.3.0
Flask==3.1.2
frozenlist==1.7.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
importlib_metadata==8.7.0
itsdangerous==2.2.0
Jinja2==3.1.6
magic-filter==1.0.12
MarkupSafe==3.0.3
multidict==6.6.4
propcache==0.3.2
pydantic==2.11.9
pydantic_core==2.33.2
python-telegram-bot==22.5
python-dotenv==1.0.1
neo4j==5.20.0
sniffio==1.3.1
typing-inspection==0.4.1
typing_extensions==4.15.0
Werkzeug==3.1.3
yarl==1.20.1
zipp==3.23.0
